# recsys-backend/docs/swagger.yaml
basePath: /
definitions:
...
  httpapi.DeviceTaskDTO:
    properties:
      deadline:
        type: string
      device_id:
        type: integer
      device_task_type_id:
        type: integer
      doc_num:
        type: string
      duration_min:
        type: integer
      id:
        type: integer
      name:
        type: string
      need_operator:
        type: boolean
      operator_id:
        type: integer
      plan_end:
        type: string
      plan_start:
        type: string
      priority_id:
        type: integer
      setup_time_min:
        type: integer
      unload_time_min:
        type: integer
      workspace_id:
        type: integer
    type: object
  httpapi.DeviceTaskRequest:
    properties:
      add_in_rec_system:
        type: boolean
      completion_mark:
        type: string
      deadline:
        type: string
      device_id:
        type: integer
      device_task_type_id:
        type: integer
      doc_num:
        type: string
      duration_min:
        type: integer
      name:
        type: string
      need_operator:
        type: boolean
      operator_id:
        type: integer
      photo_url:
        type: string
      plan_end:
        type: string
      plan_start:
        type: string
      priority_id:
        type: integer
      setup_time_min:
        type: integer
      unload_time_min:
        type: integer
    type: object
...
  service.RecomputeRequest:
    properties:
      workspace_id:
        type: integer
    type: object
  service.RecomputeResult:
    properties:
      unscheduled_ids:
        items:
          type: integer
        type: array
      updated:
        type: integer
    type: object
...
info:
  contact: {}
  description: Система рекомендаций загрузки производственного оборудования
  title: Recommendation System API
  version: "1.0"
paths:
...
  /api/plans/recompute:
    post:
      consumes:
      - application/json
      parameters:
      - description: workspace_id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.RecomputeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RecomputeResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Пересчитать рекомендации/план по workspace
      tags:
      - planning
...
  /api/workspaces/{workspaceId}/device-tasks:
    get:
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.DeviceTaskDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Список заданий по workspace
      tags:
      - device_task
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: integer
      - description: Device task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.DeviceTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать задачу оборудования
      tags:
      - device_tasks
...
swagger: "2.0"